[dsm_read_allocfile] Fixed bug that caused incorrect hosts to show when reading allocation.
This was caused by interpreting a malloc pointer as u_long but allocating the size with
sizeof(int). This works on a 32-bit system since sizeof(u_long)==size(int)==4 but on a 64-bit
system sizeof(int)==4 and sizeof(u_long)==8.

+++ dsm_read_allocfile.c
@@ -357,7 +357,7 @@ int dsm_read_allocation_file(struct alloc_list_head **alhp, u_long *nmachp,
      network on which each host has one IP address, we don't allow DSM
      to share on multiple IP addresses from the same host.  This count
      is how we check that condition. */
-  my_addresses_in_allocfile = (u_long *)calloc(num_local_addresses,sizeof(int));
+  my_addresses_in_allocfile = (u_long *)calloc(num_local_addresses,sizeof(u_long));
   if(my_addresses_in_allocfile == (u_long *)NULL) {
     perror("calloc failed");
     return(DSM_ERROR);
